+++ src/main/java/com/deare/backend/api/invite/service/SignupBenefitWriteService.java
@@ -1,25 +1,6 @@
 package com.deare.backend.api.invite.service;
 
-import com.deare.backend.api.auth.exception.AuthErrorCode;
-import com.deare.backend.domain.user.entity.User;
-import com.deare.backend.domain.user.repository.UserRepository;
-import com.deare.backend.global.common.exception.GeneralException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
+public interface SignupBenefitWriteService {
 
-@Service
-@RequiredArgsConstructor
-public class SignupBenefitWriteService {
-
-    private final UserRepository userRepository;
-    private final InviteService inviteService;
-
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public void apply(String inviteCode, Long userId) {
-        User invitee = userRepository.findById(userId)
-                .orElseThrow(() -> new GeneralException(AuthErrorCode.USER_NOT_FOUND));
-        inviteService.applySignupBenefit(inviteCode, invitee);
-    }
+    void apply(String inviteCode, Long userId);
 }
