+++ src/test/java/com/deare/backend/api/invite/InviteServiceTest.java
@@ -1,6 +1,7 @@
 package com.deare.backend.api.invite;
 
 import com.deare.backend.api.invite.service.InviteService;
+import com.deare.backend.api.invite.service.InviteServiceImpl;
 import com.deare.backend.api.invite.exception.InviteErrorCode;
 import com.deare.backend.domain.invite.entity.UserInviteCode;
 import com.deare.backend.domain.invite.repository.UserInviteCodeRepository;
@@ -22,7 +23,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DataJpaTest
-@Import({QuerydslConfig.class, InviteService.class})
+@Import({QuerydslConfig.class, InviteServiceImpl.class})
 @ActiveProfiles("test")
 class InviteServiceTest {
